Move 32/64-bit code into submodules in a `backend` module.
See the doc comment in `backend/mod.rs` for motivation on naming.
